# building-manager/technical/openapi.yaml
openapi: 3.0.3
info:
  title: Building Manager API
  description: API for the Building Manager (Upravnik Zgrade) application
  version: 1.0.0
  contact:
    name: Building Manager Team
    email: support@upravnik.app
servers:
  - url: https://api.upravnik.app/api/v1
    description: Production
  - url: https://staging-api.upravnik.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email or phone already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithProfile'

  # ============================================================================
  # BUILDINGS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
              description: Seconds until token expires
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
    UserWithProfile:
      type: object
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/Person'
    UserRole:
      type: string
      enum: [ADMIN, COMPANY_ADMIN, MANAGER, SUBSTITUTE, OWNER, TENANT]
    UserStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, DELETED]

    # -------------------------------------------------------------------------
    # Building
    # -------------------------------------------------------------------------
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        preferredName:
          type: string
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        emails:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        emergencyName:
          type: string
        emergencyPhone:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
...
    Phone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        type:
          $ref: '#/components/schemas/PhoneType'
        isPrimary:
          type: boolean
        isVerified:
          type: boolean
    PhoneType:
      type: string
      enum: [MOBILE, HOME, WORK, OTHER]
    Email:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
          format: email
        type:
          type: string
        isPrimary:
          type: boolean
        isVerified:
          type: boolean
...